requirements.txt
flask
google-api-python-client
sqlalchemy
psycopg2-binary
kafka-python
